+++ lib/reissue.rb
@@ -19,6 +19,7 @@ module Reissue
   def self.call(
     version_file:,
     changelog_file: "CHANGELOG.md",
+    retain_changelogs: false,
     segment: "patch",
     date: "Unreleased",
     changes: {},
@@ -29,7 +30,7 @@ module Reissue
     new_version = version_updater.call(segment, version_file:)
     if changelog_file
       changelog_updater = ChangelogUpdater.new(changelog_file)
-      changelog_updater.call(new_version, date:, changes:, changelog_file:, version_limit:)
+      changelog_updater.call(new_version, date:, changes:, changelog_file:, version_limit:, retain_changelogs:)
     end
     new_version
   end