+++ lib/reissue/changelog_updater.rb
@@ -18,23 +18,24 @@ module Reissue
     # @param changes [Hash] The changes for the version (default: {}).
     # @param changelog_file [String] The path to the changelog file (default: @changelog_file).
     # @param version_limit [Integer] The number of versions to keep (default: 2).
-    def call(version, date: "Unreleased", changes: {}, changelog_file: @changelog_file, version_limit: 2)
+    def call(version, date: "Unreleased", changes: {}, changelog_file: @changelog_file, version_limit: 2, retain_changelogs: false)
       update(version, date:, changes:, version_limit:)
-      write(changelog_file)
+      write(changelog_file, retain_changelogs:)
       changelog
     end
 
-    def finalize(date: Date.today, changelog_file: @changelog_file)
+    def finalize(date: Date.today, changelog_file: @changelog_file, retain_changelogs: false)
       @changelog = Parser.parse(File.read(changelog_file))
+
       # find the highest version number and if it is unreleased, update the date
-      version = changelog["versions"].max_by { |v| ::Gem::Version.new(v["version"]) }
+      version = latest_version
       version_date = version["date"]
       if version_date.nil? || version_date == "Unreleased"
         changelog["versions"].find do |v|
           v["version"] == version["version"]
         end["date"] = date
       end
-      write
+      write(changelog_file, retain_changelogs:)
       changelog
     end
 
@@ -74,8 +75,34 @@ module Reissue
     # Writes the changelog to the specified file.
     #
     # @param changelog_file [String] The path to the changelog file (default: @changelog_file).
-    def write(changelog_file = @changelog_file)
+    def write(changelog_file = @changelog_file, retain_changelogs: false)
+      if retain_changelogs
+        store_current_changelog(latest_version, retain_changelogs)
+      end
       File.write(changelog_file, to_s)
+    end
+
+    # Stores the current changelog in a file.
+    #
+    # @param version [Hash] The version information.
+    # @param path [String, Proc] The path to store the changelog or a callable that receives the version and the changelog.
+    def store_current_changelog(version, path)
+      if path
+        if path.respond_to?(:call)
+          path.call(version, to_s)
+        else
+          require "fileutils"
+          path = path.is_a?(String) ? path : "changelogs"
+          FileUtils.mkdir_p(path)
+          File.write("#{path}/#{version["version"]}.md", to_s)
+        end
+      end
+    end
+
+    private
+
+    def latest_version
+      changelog["versions"].max_by { |v| ::Gem::Version.new(v["version"]) }
     end
   end
 end