+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.3.1] - Unreleased
 
+### Added
+
+- Add `retain_changelogs` option to control how to retain the changelog files for the previous versions.
+
 ### Changed
 
 - Set the name of the `reissue:branch` argument to `branch_name` to be clearer.
 - Inject the constants in the `create` method.
+- Updated the tested Ruby version to 3.4.
 
 ## [0.3.0] - 2024-09-06